expv7.0 developing

+++ exp-v6.4/client/client_details_update.php
@@ -15,6 +15,7 @@ include_once '../config/database.php';
 include_once  '../objects/client.php';
 include_once  '../objects/device.php';
 
+
 // generate json web token
 include_once '../config/core.php';
 include_once '../libs/php-jwt-master/src/BeforeValidException.php';
@@ -27,9 +28,9 @@ use \Firebase\JWT\JWT;
 $data = json_decode(file_get_contents("php://input"));
 
 if (
-    !empty($data->jwt) && !empty($data->id) && !empty($data->payment_method) &&
-    !empty($data->name) && !empty($data->document) && !empty($data->phone) && !empty($data->area_id) && !empty($data->zone) &&
-    !empty($data->ppp_name) && !empty($data->ppp_pass) &&
+    !empty($data->jwt) && !empty($data->id) && !empty($data->mode) && !empty($data->payment_method) &&
+    !empty($data->name) && !empty($data->document) && !empty($data->phone)  && !empty($data->area_id) && !empty($data->zone) && !empty($data->expire_date) &&
+    !empty($data->disable_date) && !empty($data->ppp_name) && !empty($data->ppp_pass) &&
     !empty($data->pkg_id)
 
 ) {
@@ -47,16 +48,15 @@ if (
 
         //Assing the value in client class
         $client->id = $data->id;
-        $client->payment_method = $data->payment_method;
         $client->mode = $data->mode;
+        $client->payment_method = $data->payment_method;
         $client->name = $data->name;
         $client->phone = $data->phone;
-        $client->take_time = "0";
+        $client->take_time = $data->take_time;
         $client->area_id = $data->area_id;
         $client->zone = $data->zone;
-        $client->reg_date = date("Y-m-d H:i:s");
-        $client->expire_date = date("Y-m-d H:i:s");
-        $client->disable_date = date("Y-m-d H:i:s");
+        $client->expire_date = $data->expire_date;
+        $client->disable_date = $data->disable_date;
         $client->ppp_name = $data->ppp_name;
         $client->ppp_pass = $data->ppp_pass;
         $client->pkg_id = $data->pkg_id;
@@ -64,30 +64,32 @@ if (
         if ($client->isExistPPPname()) {
             echo json_encode(array(
                 "status" => 207,
-                "message" => "এই PPP Name টি অন্য ক্লায়েন্টের জন্য ব্যবহার করা হয়েছে।"
+                "message" => "আপনার দেওয়া PPP Name টি অন্য ক্লায়েন্টের জন্য ব্যবহার করা হয়েছে, আবার চেস্টা করুন।"
             ));
         } else if ($client->isExistPhoneToUpdate()) {
             echo json_encode(array(
                 "status" => 207,
-                "message" => "এই Phone নাম্বারটি দিয়ে একবার রেজিস্ট্রেশন হয়ে গেছে।"
+                "message" => "এই নাম্বারটি দিয়ে একবার রেজিস্ট্রেশন হয়ে গেছে।"
             ));
         } else {
 
             // ------------Mikrotik Connection Start ---------
+
             $device = new Device($db);
             $stmt = $device->get_device_url();
 
             //retrieve the table contents
             while ($row = $stmt->fetch(PDO::FETCH_ASSOC)) {
 
-                $url = $row['api_base'] . "pppCreate.php";
+                $url = $row['api_base'] . "pppUpdate.php";
                 $login_ip = $row['login_ip'];
                 $username = $row['username'];
                 $password = $row['password'];
             }
 
             $postdata = array(
-                'ppp_name' => $data->ppp_name,
+                'ppp_name' => $client->currentPppName(),
+                'ppp_new_name' => $data->ppp_name,
                 'ppp_pass' => $data->ppp_pass,
                 'pkg_id' => $data->pkg_id,
                 'mode' => $data->mode,
@@ -104,26 +106,37 @@ if (
             curl_setopt($ch, CURLOPT_HTTPHEADER, array('Content-Type: application/json'));
             $result = curl_exec($ch);
             curl_close($ch);
-
-            //Mikrotik Response handling
             $mikrotik_response = json_decode($result, true);
 
             if ($mikrotik_response['status'] == 200) {
 
-                $file_name = zuniqid() . '.jpeg';
-                $client->document =  $file_name;
+                if (file_exists('../documents/' . $data->document)) {
+                    $client->document =  $data->document;
+                    if ($client->client_details_update()) {
+                        echo json_encode(array(
+                            "status" => 200,
+                            "message" => "Details Updated Successfully."
+                        ));
+                    }
 
-                if ($client->client_details_update() && file_put_contents("../documents/" . $file_name, base64_decode($data->document))) {
-                    echo json_encode(array(
-                        "status" => 200,
-                        "message" => "Registration update Success"
-                    ));
+                } else {
+
+                    unlink('../documents/'.$client->current_document());
+                    
+                    $file_name = $data->ppp_name.'-'.$data->id.'.jpeg';
+                    $client->document =  $file_name;
+
+                    if ($client->client_details_update() && file_put_contents("../documents/" . $file_name, base64_decode($data->document))) {
+                        echo json_encode(array(
+                            "status" => 200,
+                            "message" => "Details Updated Successfully."
+                        ));
+                    }
                 }
-
             } else {
                 echo json_encode(array(
-                    "status" => $mikrotik_response['status'],
-                    "message" => $mikrotik_response['message']
+                    "status" => $data['status'],
+                    "message" => $data['message']
                 ));
             }
         }
@@ -131,8 +144,8 @@ if (
         // tell the user access denied  & show error message
         echo json_encode(array(
             "status" => 401,
-            "message" => "Access denied, error: " . $e->getMessage()
-
+            "message" => "Access denied, error: " . $e->getMessage(),
+            "error" => $e->getMessage()
         ));
     }
 } else {
